refactored IncidentForm.tsx on core to be more usable
