+++ services/core-web/src/components/common/DocumentTable.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, FC } from "react";
+import React, { FC, useEffect, useState } from "react";
 import CoreTable from "@mds/common/components/common/CoreTable";
 import {
   documentNameColumn,
@@ -7,24 +7,24 @@ import {
   uploadedByColumn,
 } from "./DocumentColumns";
 import {
-  renderTextColumn,
-  renderActionsColumn,
   ITableAction,
+  renderActionsColumn,
+  renderTextColumn,
 } from "@mds/common/components/common/CoreTableCommonColumns";
 import { some } from "lodash";
 import { closeModal, openModal } from "@mds/common/redux/actions/modalActions";
 import DocumentCompression from "./DocumentCompression";
 import { archiveMineDocuments } from "@mds/common/redux/actionCreators/mineActionCreator";
-import { connect } from "react-redux";
-import { bindActionCreators } from "redux";
+import { useDispatch, useSelector } from "react-redux";
 import { modalConfig } from "@/components/modalContent/config";
 import { Feature } from "@mds/common";
 import { SizeType } from "antd/lib/config-provider/SizeContext";
-import { ColumnType, ColumnsType } from "antd/es/table";
+import { ColumnsType } from "antd/es/table";
 import { FileOperations, MineDocument } from "@mds/common/models/documents/document";
 import {
   DeleteOutlined,
   DownloadOutlined,
+  DownOutlined,
   FileOutlined,
   InboxOutlined,
   SyncOutlined,
@@ -32,12 +32,10 @@ import {
 import { openDocument } from "../syncfusion/DocumentViewer";
 import { downloadFileFromDocumentManager } from "@common/utils/actionlessNetworkCalls";
 import { getUserAccessData } from "@mds/common/redux/selectors/authenticationSelectors";
-import { Dropdown, Button, MenuProps } from "antd";
-import { DownOutlined } from "@ant-design/icons";
+import { Button, Dropdown, MenuProps } from "antd";
 import { useFeatureFlag } from "@mds/common/providers/featureFlags/useFeatureFlag";
 import DocumentTableProps from "@mds/common/interfaces/document/documentTableProps.interface";
 
-// eslint-disable-next-line @typescript-eslint/no-shadow
 export const DocumentTable: FC<DocumentTableProps> = ({
   isViewOnly = false,
   excludedColumnKeys = [],
@@ -51,13 +49,17 @@ export const DocumentTable: FC<DocumentTableProps> = ({
   defaultSortKeys = ["upload_date", "dated", "update_timestamp"],
   view = "standard",
   canArchiveDocuments = false,
-  openModal,
-  closeModal,
   removeDocument,
-  openDocument,
   replaceAlertMessage = "The replaced file will not reviewed as part of the submission.  The new file should be in the same format as the original file.",
   ...props
 }: DocumentTableProps) => {
+  const dispatch = useDispatch();
+  const userRoles = useSelector(getUserAccessData);
+
+  const handleCloseModal = () => {
+    dispatch(closeModal());
+  };
+
   const [rowSelection, setRowSelection] = useState([]);
   const [isCompressionModal, setCompressionModal] = useState(false);
   const [isCompressionInProgress, setCompressionInProgress] = useState(false);
@@ -66,8 +68,7 @@ export const DocumentTable: FC<DocumentTableProps> = ({
 
   const allowedTableActions = {
     [FileOperations.View]: true,
-    [FileOperations.Download]: true,
-    // don't allow changes to version history where history is not shown
+    [FileOperations.Download]: true, // don't allow changes to version history where history is not shown
     [FileOperations.Replace]: !isViewOnly && showVersionHistory,
     [FileOperations.Archive]:
       !isViewOnly && canArchiveDocuments && isFeatureEnabled(Feature.MAJOR_PROJECT_ARCHIVE_FILE),
@@ -84,7 +85,7 @@ export const DocumentTable: FC<DocumentTableProps> = ({
       parsedDocs = docs.map((doc) => new MineDocument(doc));
     }
     return parsedDocs.map((doc) => {
-      doc.setAllowedActions(props.userRoles);
+      doc.setAllowedActions(userRoles);
       return doc;
     });
   };
@@ -106,68 +107,79 @@ export const DocumentTable: FC<DocumentTableProps> = ({
   const openArchiveModal = (event, docs: MineDocument[]) => {
     const mineGuid = docs[0].mine_guid;
     event.preventDefault();
-    openModal({
-      props: {
-        title: `Archive ${docs?.length > 1 ? "Multiple Files" : "File"}`,
-        closeModal: closeModal,
-        handleSubmit: async () => {
-          await props.archiveMineDocuments(
-            mineGuid,
-            docs.map((d) => d.mine_document_guid)
-          );
-          if (props.onArchivedDocuments) {
-            props.onArchivedDocuments(docs);
-          }
+    dispatch(
+      openModal({
+        props: {
+          title: `Archive ${docs?.length > 1 ? "Multiple Files" : "File"}`,
+          closeModal: handleCloseModal,
+          handleSubmit: async () => {
+            await dispatch(
+              archiveMineDocuments(
+                mineGuid,
+                docs.map((d) => d.mine_document_guid)
+              )
+            );
+            if (props.onArchivedDocuments) {
+              props.onArchivedDocuments(docs);
+            }
+          },
+          documents: docs,
         },
-        documents: docs,
-      },
-      content: modalConfig.ARCHIVE_DOCUMENT,
-    });
+        content: modalConfig.ARCHIVE_DOCUMENT,
+      })
+    );
   };
 
   const openDeleteModal = (event, docs: MineDocument[]) => {
     event.preventDefault();
-    openModal({
-      props: {
-        DocumentTable,
-        title: `Delete ${docs?.length > 1 ? "Multiple Files" : "File"}`,
-        closeModal: closeModal,
-        handleSubmit: async () => {
-          docs.forEach((record) => removeDocument(event, record.key, documentParent));
+    dispatch(
+      openModal({
+        props: {
+          DocumentTable,
+          title: `Delete ${docs?.length > 1 ? "Multiple Files" : "File"}`,
+          closeModal: handleCloseModal,
+          handleSubmit: async () => {
+            for (const doc of docs) {
+              await removeDocument(event, doc.key, documentParent);
+            }
+            handleCloseModal();
+          },
+          documents: docs,
         },
-        documents: docs,
-      },
-      content: modalConfig.DELETE_DOCUMENT,
-    });
+
+        content: modalConfig.DELETE_DOCUMENT,
+      })
+    );
   };
 
   const openReplaceModal = (event, doc: MineDocument) => {
     event.preventDefault();
-    openModal({
-      props: {
-        title: `Replace File`,
-        closeModal: closeModal,
-        handleSubmit: async (document: MineDocument) => {
-          const newDocuments = documents.map((d) =>
-            d.mine_document_guid === document.mine_document_guid ? document : d
-          );
-          setDocuments(newDocuments);
+    dispatch(
+      openModal({
+        props: {
+          title: `Replace File`,
+          closeModal: handleCloseModal,
+          handleSubmit: async (document: MineDocument) => {
+            const newDocuments = documents.map((d) =>
+              d.mine_document_guid === document.mine_document_guid ? document : d
+            );
+            setDocuments(newDocuments);
+          },
+          document: doc,
+          alertMessage: replaceAlertMessage,
         },
-        document: doc,
-        alertMessage: replaceAlertMessage,
-      },
-      content: modalConfig.REPLACE_DOCUMENT,
-    });
+        content: modalConfig.REPLACE_DOCUMENT,
+      })
+    );
   };
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const actions = [
     {
       key: "view",
       label: FileOperations.View,
       icon: <FileOutlined />,
       clickFunction: (_event, record: MineDocument) =>
-        openDocument(record.document_manager_guid, record.mine_document_guid),
+        dispatch(openDocument(record.document_manager_guid, record.mine_document_guid)),
     },
     {
       key: "download",
@@ -192,8 +204,7 @@ export const DocumentTable: FC<DocumentTableProps> = ({
     {
       key: "delete",
       label: FileOperations.Delete,
-      icon: <DeleteOutlined />,
-      // PopConfirm does not work in either the function or label field here
+      icon: <DeleteOutlined />, // PopConfirm does not work in either the function or label field here
       clickFunction: (event, record: MineDocument) => openDeleteModal(event, [record]),
     },
   ].filter((action) => allowedTableActions[action.label]);
@@ -384,20 +395,4 @@ export const DocumentTable: FC<DocumentTableProps> = ({
   );
 };
 
-const mapStateToProps = (state) => ({
-  userRoles: getUserAccessData(state),
-});
-
-// eslint-disable-next-line @typescript-eslint/no-shadow
-const mapDispatchToProps = (dispatch) =>
-  bindActionCreators(
-    {
-      openModal,
-      closeModal,
-      archiveMineDocuments,
-      openDocument,
-    },
-    dispatch
-  );
-
-export default connect(mapStateToProps, mapDispatchToProps)(DocumentTable);
+export default DocumentTable;
