+++ services/core-web/src/components/mine/Incidents/MineIncident.tsx
@@ -1,9 +1,8 @@
 import React, { FunctionComponent, useEffect, useState } from "react";
-import { bindActionCreators } from "redux";
 import { flattenObject } from "@common/utils/helpers";
-import { connect } from "react-redux";
-import { Link, withRouter, useParams, useLocation, matchPath } from "react-router-dom";
-import { change, submit, getFormSyncErrors, getFormValues, touch, isDirty } from "redux-form";
+import { useDispatch, useSelector } from "react-redux";
+import { Link, useLocation, useParams } from "react-router-dom";
+import { getFormSyncErrors, getFormValues } from "redux-form";
 import { Tag } from "antd";
 import { ArrowLeftOutlined, EnvironmentOutlined } from "@ant-design/icons";
 import { getMineIncident } from "@mds/common/redux/reducers/incidentReducer";
@@ -11,7 +10,6 @@ import {
   createMineIncident,
   fetchMineIncident,
   updateMineIncident,
-  removeDocumentFromMineIncident,
 } from "@mds/common/redux/actionCreators/incidentActionCreator";
 import { clearMineIncident } from "@mds/common/redux/actions/incidentActions";
 import * as Strings from "@mds/common/constants/strings";
@@ -21,15 +19,9 @@ import IncidentForm from "@/components/Forms/incidents/IncidentForm";
 import ScrollSideMenu from "@/components/common/ScrollSideMenu";
 import * as routes from "@/constants/routes";
 import { IMineIncident } from "@mds/common";
-import { ActionCreator } from "@mds/common/interfaces/actionCreator";
 
 export interface MineIncidentProps {
   incident: IMineIncident;
-  createMineIncident: ActionCreator<typeof createMineIncident>;
-  fetchMineIncident: ActionCreator<typeof fetchMineIncident>;
-  updateMineIncident: ActionCreator<typeof updateMineIncident>;
-  clearMineIncident(): Promise<void>;
-  removeDocumentFromMineIncident: ActionCreator<typeof removeDocumentFromMineIncident>;
   history: {
     push(): Promise<any>;
     replace(mineGuid: string, formattedValues?: any): Promise<any>;
@@ -46,7 +38,9 @@ interface IParams {
 }
 
 export const MineIncident: FunctionComponent<MineIncidentProps> = (props) => {
-  const { formValues, formErrors, incident, history } = props;
+  const dispatch = useDispatch();
+
+  const { history } = props;
   const params = useParams<IParams>();
   const { mineGuid, mineIncidentGuid = null } = params;
   const { pathname, search = null } = useLocation();
@@ -54,7 +48,11 @@ export const MineIncident: FunctionComponent<MineIncidentProps> = (props) => {
   const [isLoaded, setIsLoaded] = useState<boolean>(false);
   const [fixedTop, setIsFixedTop] = useState<boolean>(false);
 
-  const isEditPage = Boolean(matchPath(pathname, routes.EDIT_MINE_INCIDENT.route));
+  const incident = useSelector(getMineIncident);
+  const formValues = useSelector((state) => getFormValues(FORM.ADD_EDIT_INCIDENT)(state)) || {};
+  const formErrors = useSelector((state) => getFormSyncErrors(FORM.ADD_EDIT_INCIDENT)(state)) || {};
+
+  const isEditPage = !!params.mineIncidentGuid;
   const mineName = isNewIncident
     ? new URLSearchParams(search).get("mine_name")
     : incident.mine_name;
@@ -71,26 +69,17 @@ export const MineIncident: FunctionComponent<MineIncidentProps> = (props) => {
     return { url: routes.VIEW_MINE_INCIDENT, params: [mineGuid, mineIncidentGuid] };
   })();
 
-  const handleScroll = () => {
-    if (window.pageYOffset > 170 && !fixedTop) {
-      setIsFixedTop(true);
-    } else if (window.pageYOffset <= 170 && fixedTop) {
-      setIsFixedTop(false);
-    }
-  };
-
   const handleFetchData = (): Promise<void> => {
     if (mineGuid && mineIncidentGuid) {
       setIsNewIncident(false);
-      return props.fetchMineIncident(mineGuid, mineIncidentGuid);
+      return dispatch(fetchMineIncident(mineGuid, mineIncidentGuid));
     }
     return Promise.resolve();
   };
 
   const handleCreateMineIncident = (formattedValues: any) => {
     setIsLoaded(false);
-    return props
-      .createMineIncident(mineGuid, formattedValues)
+    return dispatch(createMineIncident(mineGuid, formattedValues))
       .then(({ data: { mine_guid, mine_incident_guid } }) =>
         history.replace(routes.EDIT_MINE_INCIDENT.dynamicRoute(mine_guid, mine_incident_guid))
       )
@@ -100,8 +89,7 @@ export const MineIncident: FunctionComponent<MineIncidentProps> = (props) => {
 
   const handleUpdateMineIncident = (formattedValues) => {
     setIsLoaded(false);
-    return props
-      .updateMineIncident(mineGuid, mineIncidentGuid, formattedValues)
+    return dispatch(updateMineIncident(mineGuid, mineIncidentGuid, formattedValues))
       .then(() => handleFetchData())
       .then(() => setIsLoaded(true));
   };
@@ -168,37 +156,55 @@ export const MineIncident: FunctionComponent<MineIncidentProps> = (props) => {
     isNewIncident: isNewIncident,
     incident: incident,
     handlers: {
-      deleteDocument: handleFetchData,
       handleSaveData,
+      handleFetchData: handleFetchData,
     },
   };
 
   const scrollSideMenuProps = {
     menuOptions: [
       { href: "initial-report", title: "Initial Report" },
-      { href: "incident-details", title: "Incident Details" },
+      {
+        href: "incident-details",
+        title: "Incident Details",
+      },
       { href: "documentation", title: "Documentation" },
-      { href: "final-report", title: "Final Report" },
+      {
+        href: "final-report",
+        title: "Final Report",
+      },
       { href: "ministry-follow-up", title: "Ministry Follow Up" },
-      { href: "internal-documents", title: "Internal Documents" },
+      {
+        href: "internal-documents",
+        title: "Internal Documents",
+      },
       { href: "internal-ministry-comments", title: "Comments" },
     ],
     featureUrlRoute: sideBarRoute.url.hashRoute,
     featureUrlRouteArguments: sideBarRoute.params,
   };
 
-  window.addEventListener("scroll", handleScroll);
-
   useEffect(() => {
+    const handleScroll = () => {
+      if (window.pageYOffset > 170 && !fixedTop) {
+        setIsFixedTop(true);
+      } else if (window.pageYOffset <= 170 && fixedTop) {
+        setIsFixedTop(false);
+      }
+    };
+
+    window.addEventListener("scroll", handleScroll);
+
     handleFetchData().then(() => {
       setIsLoaded(true);
-
-      return () => {
-        window.removeEventListener("scroll", handleScroll);
-        props.clearMineIncident();
-      };
     });
+
     handleScroll();
+
+    return () => {
+      window.removeEventListener("scroll", handleScroll);
+      dispatch(clearMineIncident());
+    };
   }, [pathname]);
 
   return isLoaded ? (
@@ -213,7 +219,6 @@ export const MineIncident: FunctionComponent<MineIncidentProps> = (props) => {
             &nbsp;
             <span>
               <Tag title={`Mine: ${mineName}`}>
-                {/* @ts-ignore */}
                 <Link
                   style={{ textDecoration: "none" }}
                   to={routes.MINE_GENERAL.dynamicRoute(mineGuid)}
@@ -224,7 +229,6 @@ export const MineIncident: FunctionComponent<MineIncidentProps> = (props) => {
               </Tag>
             </span>
           </h1>
-          {/* @ts-ignore */}
           <Link to={routes.MINE_INCIDENTS.dynamicRoute(mineGuid)}>
             <ArrowLeftOutlined className="padding-sm--right" />
             Back to All Incidents
@@ -246,26 +250,4 @@ export const MineIncident: FunctionComponent<MineIncidentProps> = (props) => {
   );
 };
 
-const mapStateToProps = (state) => ({
-  incident: getMineIncident(state) || {},
-  formErrors: getFormSyncErrors(FORM.ADD_EDIT_INCIDENT)(state) || {},
-  formValues: getFormValues(FORM.ADD_EDIT_INCIDENT)(state) || {},
-  formIsDirty: isDirty(FORM.ADD_EDIT_INCIDENT)(state),
-});
-
-const mapDispatchToProps = (dispatch) =>
-  bindActionCreators(
-    {
-      clearMineIncident,
-      createMineIncident,
-      fetchMineIncident,
-      updateMineIncident,
-      removeDocumentFromMineIncident,
-      submit,
-      touch,
-      change,
-    },
-    dispatch
-  );
-
-export default withRouter(connect(mapStateToProps, mapDispatchToProps)(MineIncident));
+export default MineIncident;
